requirements.txt
flask
python-docx
fpdf
pypandoc
